backend/requirements.txt
django>=5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
gunicorn>=21.2.0
